Assana-Agent-supabase/requirements.txt
livekit-agents[google]~=1.5
python-dotenv>=1.0.0
supabase>=2.3.0
